// packages/frontend/table/lib/features/filters/types.ts
/**
 * Copyright 2025 grit42 A/S. <https://grit42.com/>
 *
 * This file is part of @grit/table.
 *
 * @grit/table is free software: you can redistribute it and/or modify it
 * under the terms of the GNU General Public License as published by the Free
 * Software Foundation, either version 3 of the License, or  any later version.
 *
 * @grit/table is distributed in the hope that it will be useful, but
 * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
 * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
 * more details.
 *
 * You should have received a copy of the GNU General Public License along with
 * @grit/table. If not, see <https://www.gnu.org/licenses/>.
 */

import { GritColumnDef } from "../../types";

export interface Filter {
  id: string;
  type: string;
  operator: string;
  column: string;
  property: string;
  property_type: string;
  active: boolean;
  value: any;
}

export interface FilterOperator {
  id: string;
  name?: string;
}

export interface FilterInputProps {
  filter: Filter;
  column: GritColumnDef;
  onChange: (filter: Filter) => void;
}